requirements.txt
flask
solana
base58
PyNaCl
requests
pylibmc
